measurement-task/requirements.txt
requests
psycopg2-binary
numpy
pytz